Model: Extract the code for REGEX into on RegexUtil class
The REGEX expressions to check whether a name for a model (Name, Phone, Address, Email, Tag)
are valid or not are stored in separate files.

Since we will be dealing with Japanese characters and may have different REGEX expressions for
different models in Weeblingos, it is better to store all REGEX in one util class so that maintaince becomes easier.

Let's move all regex expressions of the 5 classes mentioned above into one RegexUtil class.
This movement does not cause any regressions.