Add ability to sort list by name
The address book does not have the ability to sort the list by name
alphabetically.

When there are a large number of contacts, the address book may look
disorganised and messy.

Implement sort method to sort the list by name in ascending alphabetical
order.

Users can thus easily find specific contacts, even if there are a large
number of contacts.