Add Matriculation Number field to 'Add' command
Matriculation Number is one of the fields that will be
contained within a Person object. It will be used for
other functions later to search for students.

This commit amends the 'Add' and 'Edit' command, 'Person'
class and other relevant classes and tests to support
the use of the Matriculation Number parameter to keep
track of students.

Since the Matriculation Number for every student is unique, it
reduces the need for UHC staff to distinguish between students
with similar names.