Add VaccinationStatus parameter in 'Add' command
Vaccination Status is one of the fields that will be contained
within a Person object. It will be used to,

*Determine if students require a vaccination if they are not already
vaccinated.
*Determine the proportion of students who have already been vaccinated.

This commit amends the 'Add' and 'Edit' command, 'Person' class and
other relevant classes and tests to support the use of the Vaccination
Status parameter.

The Vaccination Status will allow UHC staff to contact un-vaccinated
students more quickly, and inform NUS management if further reopenings
are possible and safe.