Remove Tag, add MedicalDetails parameter in 'Add'
The 'Tag' field is a legacy of the original AB3 code. It will not be
used in Vax@NUS. Medical Details is one of the fields that will be
contained within a Person object. It will be used to determine if a
student is suitable for vaccination.

This commit amends the 'Add' and 'Edit' command, 'Person' class and
other relevant classes and tests to support the use of the Medical
Details parameter. This commit also removes the 'Tag' parameter.

As life-threatening adverse drug reactions (ADRs) have been reported in
people with certain underlying medical conditions, knowing this will
help UHC staff determine who is suitable for vaccination and keep
everyone safe.