Add optional parameter 'schoolResidence'
School Residence is an optional field that will be contained within a
Person object. It will be used to determine if a student stays on
campus. It is optional as most students do not stay on campus.

This commit amends the 'Add' and 'Edit' command, 'Person' class and
other relevant classes and tests to support the use of the School
Residence parameter.

As school residences tend to be densely populated, this field will allow
UHC staff determine if an outbreak is possible, and vaccinate students
before it happens.