Add ability to add an optional remark for each contact
The address book does not have the ability to store an optional remark
for each contact.

When there are a large number of contacts, it may be difficult to
remember some of the contacts.

Implement new method to add an optional remark to each contact.

Users can thus add an optional remark to certain contacts to make it
easier to remember.