Update CliSyntax
The current CliSyntax does not support parsing for task and event objects.

Let's add additional prefixes to support parsing of task and event objects.

Updates are needed to change 'p' from the prefix of phone to the prefix of priority.
Name in our project will eventually refer to the name of the task or event instead of a person's name.