Add Not Done Functionality
Tasks marked as Done stay Done. A new task has to be created if one wants to redo the task.

This is inconvenient if one wants to mark a task as not done to highlight that the task needs to be redone.

Let's add a NotDone command.

This allows a task to be toggled between done and not done.