Add name searching via regex functionalities
Implementation only allows for simple word matching based on
substring provided. This is slightly cumbersome as it doesn't allow
find-grained searches.

For instance, it is not possible to search all person with names
starting with the alphabet 'a'

Including an option for the use of regular expressions allows the users
to have finer control over searching.

Lets create a predicate that makes use of regular expressions instead of
simple word matching.