Updated according to reviews from PR.
-Edited TypicalPersons and typicalPersonsAddressBook to include variety of date formats
-Edited DeliveryDateTest to test the .equals method
-Edited DeliveryDateTest to test for invalid delimiters and formats

Noteworthy problems that may face in the future:
DeliveryDates of persons/typical persons created in Test Cases are mostly hard-coded, as such, the test cases are not future proof since "Today's" date will keep moving forward.
This is because isValidDeliveryDate checks and makes sure deliveryDate is at least 3 days after "Today's" date.
While it is possible to code .java test files to make use of LocalDate.now() to input DeliveryDates for the test cases, I am currently uncertain if I can do the same for .json test files.
This is important as there are some .json test files that are hard-coded and used for Test Cases and so have a depency.
For now, all valid DeliveryDates used for testing is set for 2022.