Add Faculty field to 'Add' command.
Faculty is one of the fields that will be
contained within a Person object. It will be used for
other functions later to filter students by faculty.

This commit amends the 'Add' and 'Edit' command, 'Person'
class and other relevant classes and tests to support
the use of the Faculty parameter.

Knowing the faculty information will give the faculties
more information on safe reopening, allow faculties to liaise
with their students directly, reducing the burden on UHC staff.