Add multiline description, line wrapping in input
TextField used only allows a single line which can be inconvenient for
writing tasks and descriptions.

TextArea can be used to leverage on the line wrapping functionality, and
add line breaks in command input.

Let's change TextField to TextArea for the Ui text input box and replace
enter with shift+enter to introduce line breaks.