Refactor request related code
Other than some minor typos, the request class and the subclasses seem
to repeat the creation of the instance attribute endpoint.

Let's remove the duplication and modify the abstract request constructor
to take in an endpoint and save the corresponding details for further
usage by child classes via getters.

This way new child request classes need not declare an endpoint
attribute within the class and clean up the call to get headers and data
as well.