Refactor request related code
Certain parts are duplicated and repeated for no clear purposes.

Let's fix some of the violations and improve code quality in terms of
extracting out smaller methods and make minor adjustments to how request
is executed.

This is so that further inheritance and reference to the request code
can reduce the amount of lines that are previously required but now
abstracted away.