Add API to convert Meetings into Json
We need to store the meetings.

In order to store the data on meetings into json, we need to convert the classes into a more
jackson - friendly  version. Therefore it is necessary to make the class and its API.

Lets-
Make a JsonAdaptedMeeting class with conversion API
Fix checkstyle errors on some fxml files.