Add Profile Images
As a user, I would like to display profile images beside the person details in the addressbook

This is so that I can personalise and remember my contacts better. Some of the contacts would be using wordpress accounts
or have online avatars. I would want these contacts profiles to be automatically loaded. If there is a connection error, it should display an empty profiile picture.

Lets-
Add functionality to load the images from web using the Gravatar API

NOTE-
Slight implementation details to note as well as future directions
- Loading times take a while as connecting to the url takes some time
- May need to implement caching
- May need to set default profile picture and pass the choice to user in USER_PREFS in case connection fails
- May need to touch up GUI