Make dynamic addition of optional task fields in view
Optional fields of task are still rendered on the view and
would be better to appear dynamically for aesthetic
purposes. Details of a task are rendered statically in the
taskListCard fxml and should become dynamic.

Let's refactor details of the task into its own class that
is responsible for controlling the display of each task field.