Fix bug where status of task is not updated after done
Currently, a task is directly assessed and getting its status updated as complete.
However, since this does not update the observable task list and filtered task list,
this changes are not reflected unless the app restarts.

We can create a new task with the status marked as complete and replace the old task
with the new task. This is essentially a simplified edit task command. It is easy to be adapted
and can be helpful for building edit command in the next iteration.

Let's do the changes!