Adjust ListCommand message
ListCommand result is not helpful as it simply says all endpoints
listed. It does not account for the cases where there is no endpoints
saved.

Let's adjust the result output to prompt user to use show command if
there are endpoints saved, or use add command to add new endpoints.

This will make the result clearer to users.

I have also fixed the minor typos and clean up the numbering in
addCommand in order to allow users to easily copy paste the examples.
Some test cases are also fixed along the way.