Refactor JsonAdaptedEntity into individual child classes
Originally JsonAdaptedEntity is a single class that represents all types
of entities. This caused a lot of restrictions since every entity has
other unique attributes that cannot be easily captured this way,
necessitating an additional method to return the extra attributes. It
added additional complexity when restoring the entities from json while
introducing an extra method that no other parts of the program uses.

The solution here is to create the child classes, JsonAdaptedOwner and
JsonAdaptedEntity. Jackson is able to store this additional information
with a few more annotations. This makes all of the JSON adaptation of
the various entities mirror their counterparts. Also the JSON file has
been renamed to "pawbook.json" as well, aligning with the project name
along with the actual content and avoiding the potential conflicts with
the previous file format.