Ensure that referenced IDs are all valid upon import
Owners store the IDs of the dogs they own and dogs store the ID of their
owner. While the application is running it should be guaranteed that
all of these information is kept updated and always valid. But there is
no guarantee when importing from an external source.

It is therefore necessary verify that all the dog IDs referenced by an
owner are valid dogs in the list and all owner IDs reference by a dog
are valid owners. There is also a need to verify that these links are
mutual.