Refactor rendering of taskfields onto view, enable assert
Each of the task fields in the taskCardDetails have individual
methods and have a lot of duplicate code.

Using a command-like pattern can be more easily extensible
for the rendering of each field. It also maintains the
responsibility of knowing the values of each field within
Task.

Let's use Task to return the compulsory and optional fields
and their values to the TaskCardDetails and render them
according to these 2 variables instead of by fields.