Include Meeting Display In Order of Time
Add List of Meetings displayed on right side of screen

Testing to be added on next commit.

Reason of Implementation of Meeting under Person:

The Meeting List is implemented via TreeMap with the key being
Meeting and the value being Person.

When we edit the person contacts, the meeting key is still easily
accessible as the meeting does not change despite the person changing.

If we edit the meeting, as meetings can only be added onto a particular
person, the prechanged and postchanged meetings are both accessible.

If Person is implemented under Meeting, then everytime we edit a person,
we would have to iterate through the meeting list to change the person.

Hence, even though this creates implementation restrictions on our code,
we've decided to implement the Meeting as an attribute to Person for the
time being.

Side Note:

Optional obejcts are opened by get() rather than a safer function is
due to method mapping in the Optional has lead to bugs in various
occasions.