Refactor Tags into a UniqueTagList
Each Task object needs their own set of Tag objects and they
might not be unique. It would be more Object-Oriented to
allow a Task consist of some Tags that are shared among all
the Tasks, as this allows Tag to be a separate entity without
needing to be instantiated for a Task.

Let's create a UniqueTagList under Planner that exposes an
observableList for Ui to watch, and update this
UniqueTagList whenever there is a change.