Handle Meeting clashes
there can be multiple meetings in the model that clash. This is not reflective of the user
scenario as it should only keep track of all meetings that the user is able to attend.

meeting is not added, and a helpful error message displaying the clashing  meetings will be shown.

Lets
- Throw a CommandTypeException a clash is detected
- Make setMeeting and add in methods in unique meeting list
throw Runtime exception when a meeting is added.
-UniqueMeetingList will now only be allowed to keep meetings that do not clash