Fix table view
Display up to rows in the timetable, no empty rows should be displayed

Do not seem to have nice way to set the bind the cell size.

Therefore just bind the table size to cell size + column header size.
The fix the cell size to make it fit the fullscreen.