Modify upcoming event sorting algorithm
Changed to adhere more closely to how upcoming birthdays are sorted, i.e. overdue birthdays/events are always sorted to the back. There is an additional logic here: if event is done, this should also be checked off the upcoming event list, and so is funnelled to the back as well.