Add functionality to process the new attribute of Date Of Birth
Currently, we only have the bare class that contains the information about Date Of Birth

This obviously cannot be used in production, and the class is currently unused.

Let's,
* Update the customer to include the new attribute
* Update the AddCommand and EditCommand classes and related classes to process the new attribute
* Update general parsing classes and the appropriate storage classes to handle the new attribute
* Fix any resulting test code that is unable to handle the new attribute

The reason we do this all at once is so that we can maintain perfect test coverage. Note that we also
update test cases to test for this attribute specifically

{any other relevant info}