Link viewing date to calendar view
Viewing date only filters the list but not show the date on
the calendar when the calendar is so prominent.

Linking the viewing date to the calendar view will be more
intuitive and convenient for the user.

Let's make an observable calendar date which allows the
calendar panel to receive updates from the view day
command, and update the calendar view accordingly.