Fix bug where delete will fail when entity not visible
The bug occurred due to delete commands manually matching the target ID
against the filtered list, resulting in the commands being unable to
find the correct entity even when it truly exists. The fix here is to
make use of the standardised methods from model to perform the matching
and retrieval, avoiding raw use of the underlying list of entities.