Fix(Parser): Update parser to deal with preamble better
The parser now deals with non-integer and integer input separately. This
allows for better error messages to be raised to the user to explain why
the entered preamble used in index values may be failing.

This commit also separates non-integer input to throw an
illegalargumentexception instead of a parseexception, which is now
reserved for integer input that fails (<=0, > range of list, etc)