Modifies PersonInMeetingPredicate such that only persons in meeting are shown when showm command is called.
Previously, showm would show both the persons in the meeting as well as persons in groups that are in the meeting.
After a team discussion, it was decided that persons in groups that are in the meeting will not be shown, and that only the persons in the meeting should be shown.