Remove scroll bar in tag panel
Tag panel uses ListView whose slider resizes according to
the elements it contains. This causes a glitchy resizing issue
when the elements sizes are too different. This is a known
bug of JavaFX and has no clear solution.

The issue can be solved by hiding the slider such that users
cannot access the issue at all.